% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_quality.R
\name{quality_report}
\alias{quality_report}
\title{Construire un rapport de qualite pour un dataset}
\usage{
quality_report(dataset)
}
\arguments{
\item{dataset}{Un \code{mecore::me_dataset}.}
}
\value{
Un \code{me_quality_report}.
}
\description{
Construire un rapport de qualite pour un dataset
}
